New Debug Tab
In preparation for breakpoints being added, a new tab called "Debug" is
pretty much ready to go. It contains the register screen of the old
Breakpoints tab, along with a list of all threads active on the Wii U,
with options to pause and view/change registers. Will work in
conjunction with actual breakpoints when that's complete.